chore: auth 경로 삭제
