+++ src/app/(routes)/(root)/initial/common-nickname/page.tsx
