+++ src/app/(routes)/(root)/initial/layout.tsx
